fix: correct initial issues (#1)
* fix: run install as sudo

* fix: make install executable + double check ci

* fix: load secrets from file

* fix: test for secret file

* fix: lint issues

* ci: lint playbook

* fixup: leftover character improved secrets

* ci: add shellcheck

* refactor: move playbook to root

* refactor: yaml lint

* refactor!: correct system-core -> system_core

* fix: give root

* fix: add sudo

* fix: set user shell

* fix: add user .ssh

* docs: add CI badge

* feat: use my dotfiles by default

* ci: load custom values in ci

* fixup: remove sudo line

* fix: Ensure GitHub is added to known hosts for ssh

* fixup: fix knownhosts

* fix: use https with git module

* fix: add sudo

* fix: ensure dotfile installer is executable

* fix: git config args

* fix: install gh args

* fix: yq tag

* fix: neovim tag

* fix: bat install

* fix: indent

* fix: bats-core ssh based download from git

* fix: create tmp dir

* fix: use make module to install

* fix: improve go version check

* fix: ignore errors

* fix: go

* fix: nvm to https

* fix: node install

* fix: convert source to '.'

* fix: source nvm when setting version

+++ install.sh
@@ -10,12 +10,19 @@ sudo apt-add-repository --yes --update ppa:ansible/ansible
 sudo apt install ansible
 
 # To keep the repo clean, I load my defaults from a separate untracked file.
-if ./defaults/.main.yml; then
+if [ -f ./defaults/.main.yml ]; then
+  echo "Overwriting variables with secret defaults..."
   mv ./defaults/.main.yml ./defaults/main.yml
 fi
+## In CI, I load from a JSON file loaded in the environment.
+if [ -f ./defaults/.main.json ]; then
+  echo "Overwriting variables with secret defaults..."
+  mv ./defaults/.main.json ./defaults/main.yml
+fi
+
 
 echo "Install Ansible Galaxy requirements"
 ansible-galaxy install -r requirements.yml
 
 echo "Running playbook"
-ansible-playbook ./playbooks/linux-dev.yml
+sudo ansible-playbook ./linux-dev.yml --extra-vars "@./defaults/main.yml"
